+++ src/commands/includes/moveParentToWaitIfNeeded.lua
@@ -13,7 +13,7 @@ local function moveParentToWaitIfNeeded(parentQueueKey, parentDependenciesKey,
                                         parentKey, parentId, timestamp)
     local isParentActive = rcall("ZSCORE",
                                  parentQueueKey .. ":waiting-children", parentId)
-    if rcall("SCARD", parentDependenciesKey) == 0 and isParentActive then
+    if isParentActive and rcall("SCARD", parentDependenciesKey) == 0 then
         rcall("ZREM", parentQueueKey .. ":waiting-children", parentId)
         local parentWaitKey = parentQueueKey .. ":wait"
         local parentPausedKey = parentQueueKey .. ":paused"